Add syscall

+++ Assignment5/linux-stable/include/linux/mm_types_task.h
@@ -1,4 +1,3 @@
-/* SPDX-License-Identifier: GPL-2.0 */
 #ifndef _LINUX_MM_TYPES_TASK_H
 #define _LINUX_MM_TYPES_TASK_H
 
@@ -32,7 +31,7 @@
 #define VMACACHE_MASK (VMACACHE_SIZE - 1)
 
 struct vmacache {
-	u64 seqnum;
+	u32 seqnum;
 	struct vm_area_struct *vmas[VMACACHE_SIZE];
 };
